+++ NEWS.md
@@ -1,4 +1,8 @@
-# admtools (development version)
+# admtools 0.3.1
+
+* more options for sedimentation rate generator `sed_rate_from_matrix`
+
+* abstracted exctraction of tie points
 
 # admtools 0.3.0